+++ example/call_graph/__init__.py
@@ -1,0 +1,1 @@
+"""Generate call graph for a playbook execution"""